UPSTREAM 1715: Add server name for the Machine InternalDNS
For a node to join the cluster, a CSR generated by kubelet with
the node name must be approved. The approval happens if the value
of the NodeInternalDNS entry on the Machine matches the node name.
When using legacy cloud provider the value of that node name is the
Server name. This commit ensures the nodes handled with legacy cloud
provides get approved by adding the server name to the list of Machine
addresses.

(cherry picked from commit de8aaa481fef13c1d2da4b53d52511155a7a810c)