api: remove CEL validations for `AdditionalBlockDevices`
I've put that into a separate commit so we can re-add it later.
It also serves as documentation for the work we have done to add CEL.

However and unfortunately, we don't have CEL tests in place now so we
decided to not have CEL validations for now; add them later and then
we'll be able to revert this commit.

(cherry picked from commit 71fa48fa3fac6f3c30a4955357d8da11c970da76)