api: Add ServerGroupFilter
Allow users to request Server Groups by either name or ID. This aligns
us with the pattern of filters used for other resources such as Security
Groups and Ports.

Additionally `ReferencedResources` field is added to `BastionStatus` and
`OpenStackMachineStatus` as a container to hold all the IDs of resources
computed by the controllers. `ServerGroupID` is the first field added
there.

Co-Authored-By: Emilien Macchi <emilien@redhat.com>
Co-Authored-By: Michał Dulko <mdulko@redhat.com>
Co-Authored-By: Stephen Finucane <stephenfin@redhat.com>