conversion-gen: Fix package name generation
When conversion-gen has to generate a dependency between 2 input
packages it uses the value passed to --input-dirs as the package name in
the generated import. This does not work if that value is a directory
name. If you specify a package name here instead it works correctly.

We also stop trying to generate conversions for versions which don't
need them.