Remove logger from scope
Scope encapsulates an initialised OpenStack client which can be safely
shared between reconciles which share the same credentials. The logger
is initialised with context specific to an individual reconcile, so it
cannot be shared between reconciles.

To avoid code churn we preserve the previous interface of Scope with the
creation of scope.WithLogger, which changes the initialisation to ensure
that the returned logger always comes from the current reconcile even if
the scope was cached.

While touching logging in tests, we also take the opportunity to replace
uses of logr.Discard with GinkgoLogr or logr/testr as appropriate, so
test logs are now captured.