Split v1alpha6&v1alpha7 conversion into multiple files
This is 100% code motion with no code changes whatsoever.

conversion.go has become unwieldy, making changes to it harder to
understand. This simply splits it into several files based on the
corresponding _types.go.

We don't touch v1alpha5 as it's about to be deleted. It can stay a mess.