Bastion is enabled by default if specified
Bastion.Enabled is a wart. Ideally it would not be required, but because
of limitations in how we delete the bastion we require an intermediate
'disabled' step before removal. Ultimately we intend to remove this
limitation.

Eventually we would like to be able to ignore enabled entirely. e.g. To
create a Bastion just specify it:

```yaml
spec:
  bastion:
    spec:
      ...
    floatingIP: x.x.x.x
```

and to delete it just remove the bastion field.

Right now with enabled defaulting to `false`, doing the above will not
result in the creation of a bastion, because enabled must be explicitly
set to true.

Paving the way for the eventual deprecation of Bastion.Enabled, we
change the default value of enabled to be true so the above does today
what we eventually want it to do. This is also generally more intuitive:
why would you include a bastion and a spec if you didn't want to create
a bastion? Having to also set enabled to true is currently a trip
hazard.

Until we resolve the limitations of bastion deletion, though, we still
need to be able to disable the bastion. For this case enabled can be
explicitly set to false.

In the future when we remove the requirement to disable the bastion
before deletion the user can simply ignore Bastion.Enabled, which will
continue to work but without the limitations.